Rename verify to verifier.
